Modified transactions with positions and buy/sell

+++ internal/models/transaction.go
@@ -7,7 +7,7 @@ type Transaction struct {
     UserID   int64 `json:"user_id" binding:"required"`
     AssetID int64 `json:"player_id" binding:"required"`
 	Type string `json:"type" binding:"required`
-    Quantity float64 `json:"quantity" binding:"required"`
+    Quantity int `json:"quantity" binding:"required"`
 	Price float64  `json:"price" binding:"required"`
 	Timestamp time.Time `json:"timestamp" binding:"required"`
